split out validate and add multi-query support
