+++ bergson/magic/cli.py
@@ -1,17 +1,13 @@
-import json
 import os
-import random
 import shutil
 import time
 from dataclasses import asdict
 from pathlib import Path
-from typing import Callable
 
 import torch
 import torch.distributed as dist
-import torchopt
 from datasets import Dataset
-from scipy.stats import describe, pearsonr, spearmanr
+from scipy.stats import describe
 from simple_parsing import ArgumentParser
 from torch.distributed._functional_collectives import (
     all_reduce as differentiable_all_reduce,
@@ -19,10 +15,9 @@ from torch.distributed._functional_collectives import (
 from torch.distributed._functional_collectives import (
     wait_tensor,
 )
-from torch.distributed.tensor import init_device_mesh
 from torchopt.pytree import tree_iter
 from tqdm import tqdm
-from transformers import AutoModelForCausalLM, AutoTokenizer
+from transformers import AutoTokenizer
 from transformers.utils.logging import (
     disable_progress_bar as hf_disable_pbar,
 )
@@ -30,26 +25,19 @@ from transformers.utils.logging import (
     set_verbosity_error as hf_set_verbosity_error,
 )
 
-from ..config.config import ScoreConfig, TrainingConfig, ValidationConfig
-from ..config.config_io import load_subconfig, save_run_config
-from ..data import load_scores
+from ..config.config import TrainingConfig, ValidationConfig
+from ..config.config_io import save_run_config
 from ..distributed import grad_tree, launch_distributed_run
-from ..utils.csv_writer import CSVWriter
 from ..utils.load_from_optimizer import (
     save_second_moments_as_optimizer_pt,
 )
 from ..utils.logging import wandb_log_fn
 from ..utils.utils import get_device, get_device_index
-from ..utils.worker_utils import (
-    setup_data_pipeline,
-    setup_model_and_peft,
-)
+from ..utils.worker_utils import setup_data_pipeline
+from ..validate import load_attribution_scores, validate_scores
 from .config import MagicConfig
-from .data_stream import DataStream
-from .dtensor_patch import apply_dtensor_patch
-from .fsdp import simple_fsdp
-from .optim import muon
-from .trainer import BackwardState, Trainer, TrainerState
+from .data_stream import DataStream, pad_dataset_to_batch_size
+from .trainer import BackwardState, TrainerState, prepare_trainer
 
 
 def compute_query_gradients(
@@ -115,64 +103,6 @@ def compute_query_gradients(
     return grad_accum, float(loss_accum)
 
 
-def prepare_trainer(cfg: TrainingConfig, rank: int, schedule: Callable):
-    """Prepare the model, optimizer, and trainer for training."""
-    model, target_modules = setup_model_and_peft(
-        cfg,
-        attn_implementation="eager",
-        apply_fsdp=False,
-    )
-    model.to(get_device(rank))  # type: ignore[reportArgumentType]
-
-    if target_modules:
-        # Only train the PEFT adapter parameters
-        model.requires_grad_(False)
-        for name in target_modules:
-            module = model.get_submodule(name)
-            module.requires_grad_(True)
-    else:
-        model.requires_grad_(True)
-
-    if cfg.grad_checkpointing:
-        model.gradient_checkpointing_enable(  # type: ignore[attr-defined]
-            gradient_checkpointing_kwargs=dict(use_reentrant=False),
-        )
-
-    if cfg.fsdp and dist.is_initialized():
-        apply_dtensor_patch()
-        mesh = init_device_mesh("cuda", (dist.get_world_size(),))
-        with mesh:
-            model = simple_fsdp(model)
-
-    match cfg.optimizer:
-        case "adamw":
-            opt = torchopt.adamw(
-                schedule,
-                betas=(cfg.adam_beta1, cfg.adam_beta2),
-                eps_root=cfg.eps_root,
-                weight_decay=cfg.weight_decay,
-            )
-        case "muon":
-            opt = muon(
-                schedule,
-                momentum=cfg.adam_beta1,
-                adamw_betas=(cfg.adam_beta1, cfg.adam_beta2),
-                adamw_eps_root=cfg.eps_root,
-                weight_decay=cfg.weight_decay,
-            )
-        case "sgd":
-            opt = torchopt.sgd(
-                schedule,
-                momentum=cfg.adam_beta1,
-                weight_decay=cfg.weight_decay,
-            )
-        case other:
-            raise ValueError(f"Unsupported optimizer: {other}")
-
-    trainer, fwd_state = Trainer.initialize(model, opt)
-    return trainer, fwd_state, model
-
-
 def attach_doc_ids_if_missing(dataset: Dataset) -> Dataset:
     """Ensure the dataset has a ``doc_ids`` column.
 
@@ -197,63 +127,6 @@ def attach_doc_ids_if_missing(dataset: Dataset) -> Dataset:
         with_indices=True,
         desc="Attaching doc_ids",
     )
-
-
-def pad_dataset_to_batch_size(
-    dataset: Dataset,
-    batch_size: int,
-    num_docs: int,
-    label: str,
-    global_rank: int,
-) -> tuple[Dataset, int, int, int]:
-    """Pad dataset to be divisible by batch_size by repeating the last example.
-
-    Returns (padded_dataset, num_docs, pad_count, weight_pad_count).
-
-    `pad_count` is the number of rows appended to the dataset (0 if unchanged).
-    `weight_pad_count` is the number of trailing entries of a *1D* per-doc
-    weight tensor that should be zeroed to silence the pad rows' training
-    contribution.
-
-    - If the dataset has a "doc_ids" column, `.select(total - 1, ...)` copies
-      the last doc's doc_ids into every pad row. Zeroing the last `pad_count`
-      entries of a weights-indexed-by-doc_id tensor would silence real docs,
-      so we instead route pad rows to a fresh synthetic doc id (=num_docs),
-      bump num_docs by 1, and set `weight_pad_count = 1`.
-    - Otherwise rows are self-identified docs: num_docs becomes the padded
-      length and `weight_pad_count = pad_count` zeros the pad rows directly.
-
-    In per-token (2D) mode callers should zero `weights[-pad_count:]` instead
-    — `weight_pad_count` applies only to 1D per-doc weights.
-    """
-    remainder = len(dataset) % batch_size
-    if not remainder:
-        return dataset, num_docs, 0, 0
-
-    pad_count = batch_size - remainder
-    total = len(dataset)
-    pad_indices = list(range(total)) + [total - 1] * pad_count
-    dataset = dataset.select(pad_indices)
-
-    if "doc_ids" in dataset.column_names:
-        synthetic_doc_id = num_docs
-        new_doc_ids = [
-            row if i < total else [synthetic_doc_id] * len(row)
-            for i, row in enumerate(dataset["doc_ids"])
-        ]
-        dataset = dataset.remove_columns("doc_ids").add_column("doc_ids", new_doc_ids)
-        num_docs += 1
-        weight_pad_count = 1
-    else:
-        num_docs = len(dataset)
-        weight_pad_count = pad_count
-
-    if global_rank == 0:
-        print(
-            f"{label}: padded {pad_count}/{total} examples "
-            f"(weight=0) to fill last batch"
-        )
-    return dataset, num_docs, pad_count, weight_pad_count
 
 
 def worker(
@@ -403,6 +276,7 @@ def worker(
         fwd_state, model, query_stream, run_cfg.query_method, run_cfg.fsdp
     )
 
+    multi_query = False
     if not score_path:
         # Sanity check
         if not isinstance(run_cfg, MagicConfig):
@@ -452,11 +326,8 @@ def worker(
             score_path = os.path.join(run_cfg.run_path, "scores.pt")
             torch.save(scores, score_path)
             print(f"Saved attribution scores to {score_path}")
-    elif os.path.isdir(score_path):
-        scores = torch.from_numpy(load_scores(Path(score_path))[:])
-        score_cfg = load_subconfig(score_path, "score_cfg", ScoreConfig)
-        if score_cfg is not None and score_cfg.higher_is_better:
-            scores = -scores
+    elif os.path.isdir(score_path) or score_path.endswith(".npy"):
+        scores, multi_query = load_attribution_scores(score_path)
     else:
         scores = torch.load(score_path, map_location="cpu")
 
@@ -476,135 +347,24 @@ def worker(
     if getattr(run_cfg, "skip_validation", False):
         return
 
-    # Validate attribution scores via leave-subset-out retraining
-    diffs = []
-    score_sums = []
-
-    assert scores.ndim == 1 or scores.shape[1] == 1
-    scores = scores.flatten()
-
-    if run_cfg.exclude_zero_scores:
-        valid_indices = torch.nonzero(scores != 0, as_tuple=True)[0]
-    else:
-        valid_indices = torch.arange(len(scores))
-
-    if run_cfg.subset_strategy == "random":
-        rng = torch.Generator().manual_seed(run_cfg.seed)
-        if run_cfg.subset_fraction > 0:
-            # Draw potentially overlapping samples
-            subset_size = max(1, round(run_cfg.subset_fraction * len(valid_indices)))
-
-            subsets = [
-                valid_indices[
-                    torch.randperm(len(valid_indices), generator=rng)[:subset_size]
-                ]
-                for _ in range(run_cfg.num_subsets)
-            ]
-        else:
-            # Draw non-overlapping samples
-            perm = valid_indices[torch.randperm(len(valid_indices), generator=rng)]
-
-            # Shuffle the order of the subsets so that the estimate of
-            # correlation on the progress bar is unbiased. This does not change
-            # the final correlation since all subsets are eventually evaluated,
-            # but prevents the early subsets from being biased towards higher
-            # or lower scores.
-            subsets = list(perm.chunk(run_cfg.num_subsets))
-            rng = random.Random(run_cfg.seed)
-            rng.shuffle(subsets)
-    else:
-        raise ValueError(f"Unknown subset strategy: {run_cfg.subset_strategy}")
-
-    csv_path = os.path.join(run_cfg.run_path, "validation.csv")
-    val_csv_writer = CSVWriter(
-        csv_path,
-        columns=["subset", "diff", "score_sum"],
-        enabled=global_rank == 0,
+    validate_scores(
+        run_cfg,
+        scores,
+        multi_query,
+        global_rank=global_rank,
+        rank=rank,
+        world_size=world_size,
+        schedule=schedule,
+        stream=stream,
+        query_stream=query_stream,
+        fwd_state=fwd_state,
+        model=model,
+        baseline=baseline,
+        num_query_docs=num_query_docs,
+        query_weight_pad_count=query_weight_pad_count,
+        pad_count=pad_count,
+        weight_pad_count=weight_pad_count,
     )
-
-    # Disable annoying repetitive model loading messages, even on rank 0
-    hf_disable_pbar()
-    hf_set_verbosity_error()
-
-    # Optionally persist each retrained model for later attribution queries.
-    save_models = getattr(run_cfg, "save_retrained_models", False)
-    retrained_tokenizer = None
-    if save_models and global_rank == 0:
-        retrained_tokenizer = AutoTokenizer.from_pretrained(
-            run_cfg.tokenizer or run_cfg.model
-        )
-
-    pbar = tqdm(subsets, desc="Validating", disable=global_rank != 0)
-    for i, subset in enumerate(pbar):
-        trainer, fwd_state, model = prepare_trainer(run_cfg, rank, schedule)
-        fwd_state.detach_()
-
-        stream.weights.fill_(1.0)
-        if pad_count:
-            if stream.weights.ndim == 1:
-                stream.weights.data[-weight_pad_count:] = 0.0
-            else:
-                stream.weights.data[-pad_count:] = 0.0
-        stream.weights[subset] = 0.0
-
-        for x in stream:
-            fwd_state = trainer.step(fwd_state, x, inplace=True, fsdp=run_cfg.fsdp)
-
-        if save_models and global_rank == 0:
-            out_dir = os.path.join(run_cfg.run_path, "retrained", f"subset_{i}")
-            os.makedirs(out_dir, exist_ok=True)
-            with fwd_state.activate(model), torch.no_grad():
-                model.save_pretrained(out_dir, safe_serialization=True)
-            if retrained_tokenizer is not None:
-                retrained_tokenizer.save_pretrained(out_dir)
-
-        with fwd_state.activate(model), torch.no_grad():
-            loss = torch.tensor(0.0, device=stream.weights.device)
-            for batch in query_stream:
-                del batch["example_weight"]
-
-                loss += model(**batch).loss.detach() / len(query_stream)
-
-        if world_size > 1:
-            dist.all_reduce(loss, op=dist.ReduceOp.AVG)
-
-        diff = baseline - loss.item()
-        score_sum = scores[subset].sum().item()
-        val_csv_writer.writerow(i, diff, score_sum)
-
-        if global_rank == 0:
-            diffs.append(diff)
-            score_sums.append(score_sum)
-
-            if len(diffs) >= 2:
-                sp = spearmanr(diffs, score_sums)
-                pe = pearsonr(diffs, score_sums)
-                pbar.set_postfix({"rho": sp.statistic, "r": pe.statistic})
-            else:
-                pbar.set_postfix({"rho": "n/a", "r": "n/a"})
-
-    val_csv_writer.close()
-    if global_rank == 0:
-        sp = spearmanr(diffs, score_sums)
-        pe = pearsonr(diffs, score_sums)
-        print(f"Final Spearman correlation: {sp.statistic:.4f} (p={sp.pvalue:.2e})")
-        print(f"Final Pearson correlation:  {pe.statistic:.4f} (p={pe.pvalue:.2e})")
-        print(f"Saved validation data to {csv_path}")
-
-        summary_csv_writer = CSVWriter(
-            os.path.join(run_cfg.run_path, "summary.csv"),
-            columns=[
-                "spearman_corr",
-                "spearman_p",
-                "pearson_corr",
-                "pearson_p",
-                "N",
-                "baseline_loss",
-            ],
-        )
-        summary_csv_writer.writerow(
-            sp.statistic, sp.pvalue, pe.statistic, pe.pvalue, len(subsets), baseline
-        )
 
 
 def run_magic(run_cfg: TrainingConfig, *, score_path: str = ""):
@@ -653,161 +413,6 @@ def run_magic(run_cfg: TrainingConfig, *, score_path: str = ""):
         [train_ds, query_ds, train_n, query_n, run_cfg, score_path],
         run_cfg.distributed,
     )
-
-
-def evaluate_retrained(
-    run_cfg: ValidationConfig,
-    retrained_dir: str,
-    *,
-    score_path: str = "",
-):
-    """Evaluate a bank of pre-saved leave-k-out models on a query, no retraining.
-
-    Reads models written by an earlier run with
-    ``save_retrained_models=true`` and evaluates attribution scores. No training
-    happens so evaluation is cheap.
-    """
-    assert score_path, "evaluate_retrained requires precomputed --scores"
-    src = Path(retrained_dir)
-    models_root = src / "retrained"
-    subsets_path = src / "subsets.json"
-    if not subsets_path.exists():
-        raise FileNotFoundError(
-            f"{subsets_path} not found; retrained_dir must point at a run "
-            f"directory written with save_retrained_models=true"
-        )
-
-    run_path = Path(run_cfg.run_path)
-    run_path.mkdir(parents=True, exist_ok=True)
-    save_run_config(run_cfg, run_path)
-
-    # Row i of subsets.json lists the doc ids left out of retrained/subset_i, so
-    # scores[subsets[i]] is the summed attribution of exactly what model i drops.
-    with open(subsets_path) as f:
-        subset_lists = json.load(f)
-    subsets = [torch.tensor(s, dtype=torch.long) for s in subset_lists]
-
-    # Load per-query attribution scores (mirrors run_magic's score loading).
-    if os.path.isdir(score_path):
-        scores = torch.from_numpy(load_scores(Path(score_path))[:])
-        score_cfg = load_subconfig(score_path, "score_cfg", ScoreConfig)
-        if score_cfg is not None and score_cfg.higher_is_better:
-            scores = -scores
-    else:
-        scores = torch.load(score_path, map_location="cpu")
-    assert (
-        scores.ndim == 1 or scores.shape[1] == 1
-    ), "evaluate_retrained expects per-doc (1D) scores"
-    scores = scores.flatten()
-
-    max_idx = max((int(s.max()) for s in subsets if len(s)), default=-1)
-    if max_idx >= len(scores):
-        raise ValueError(
-            f"subsets.json references doc id {max_idx} but scores has only "
-            f"{len(scores)} entries -- the query must be scored against the same "
-            f"training set and seed used to build the retrained bank"
-        )
-
-    # Build the query stream on a single device (no distributed training here).
-    device = get_device(0)
-    query_ds, query_n = setup_data_pipeline(run_cfg, run_cfg.query)
-    query_ds, query_n, q_pad, q_weight_pad = pad_dataset_to_batch_size(
-        query_ds, run_cfg.batch_size, query_n, "Query", 0
-    )
-    if len(query_ds) < run_cfg.batch_size:
-        raise ValueError(
-            f"Query dataset has {len(query_ds)} examples, fewer than "
-            f"batch_size={run_cfg.batch_size}. Use a larger query split or "
-            f"smaller batch_size."
-        )
-    query_stream = DataStream(
-        query_ds,
-        run_cfg.batch_size,
-        device=device,
-        input_key=run_cfg.query.prompt_column,
-        weight_shape=(query_n,),
-    )
-    if q_pad:
-        query_stream.weights.data[-q_weight_pad:] = 0.0
-
-    hf_disable_pbar()
-    hf_set_verbosity_error()
-
-    def query_loss(model: torch.nn.Module) -> float:
-        """Mean query loss for an already-loaded model."""
-        model.eval()
-        total = torch.tensor(0.0, device=device)
-        with torch.no_grad():
-            for batch in query_stream:
-                del batch["example_weight"]
-                total += model(**batch).loss.detach() / len(query_stream)
-        return float(total)
-
-    # The bank's no-leave-out model (retrained/base) gives the query baseline;
-    # absent it, baseline stays 0 (correlation-safe).
-    base_dir = models_root / "base"
-    baseline = 0.0
-    if base_dir.exists():
-        base = AutoModelForCausalLM.from_pretrained(
-            base_dir, dtype=torch.float32, attn_implementation="eager"
-        ).to(
-            device  # type: ignore
-        )
-        baseline = query_loss(base)
-        del base
-        print(f"Baseline query loss (no leave-out) from {base_dir}: {baseline}")
-
-    csv_path = os.path.join(run_cfg.run_path, "validation.csv")
-    val_csv_writer = CSVWriter(csv_path, columns=["subset", "diff", "score_sum"])
-
-    diffs: list[float] = []
-    score_sums: list[float] = []
-    pbar = tqdm(range(len(subsets)), desc="Evaluating")
-    for i in pbar:
-        model_dir = models_root / f"subset_{i}"
-        model = AutoModelForCausalLM.from_pretrained(
-            model_dir, dtype=torch.float32, attn_implementation="eager"
-        ).to(
-            device  # type: ignore
-        )
-        loss = query_loss(model)
-        del model
-
-        diff = baseline - loss
-        score_sum = scores[subsets[i]].sum().item()
-        val_csv_writer.writerow(i, diff, score_sum)
-
-        diffs.append(diff)
-        score_sums.append(score_sum)
-        if len(diffs) >= 2:
-            sp = spearmanr(diffs, score_sums)
-            pe = pearsonr(diffs, score_sums)
-            pbar.set_postfix({"rho": sp.statistic, "r": pe.statistic})
-
-    val_csv_writer.close()
-    print(f"Saved validation data to {csv_path}")
-
-    sp = spearmanr(diffs, score_sums)
-    pe = pearsonr(diffs, score_sums)
-    print(
-        f"Spearman: {sp.statistic:.4f} (p={sp.pvalue:.2e})  "
-        f"Pearson: {pe.statistic:.4f} (p={pe.pvalue:.2e})"
-    )
-    summary_csv_writer = CSVWriter(
-        os.path.join(run_cfg.run_path, "summary.csv"),
-        columns=[
-            "spearman_corr",
-            "spearman_p",
-            "pearson_corr",
-            "pearson_p",
-            "N",
-            "baseline_loss",
-        ],
-    )
-    summary_csv_writer.writerow(
-        sp.statistic, sp.pvalue, pe.statistic, pe.pvalue, len(subsets), baseline
-    )
-    summary_csv_writer.close()
 
 
 def main():
